Add TabbedPage TabBarIsVisible In Shared & iOS
Added new attached property, I've walked through shells implementation and copied the principle in the tabbed renderer. When the value toggles the tab bar should show and hide on the same page & when navigating the pages should set the tab bar is hidden property, although I'm not as sure they will!